# Optimize small and medium int in1d operations