--- a/benchmarks/in1d.py
+++ b/benchmarks/in1d.py
@@ -42,8 +42,8 @@ def create_parser():
     parser = argparse.ArgumentParser(description="Measure the performance of in1d: c = ak.in1d(a, b)")
     parser.add_argument('hostname', help='Hostname of arkouda server')
     parser.add_argument('port', type=int, help='Port of arkouda server')
-    parser.add_argument('-n', '--size', type=int, default=10**7, help='Problem size: length of array a')
-    parser.add_argument('-t', '--trials', type=int, default=1, help='Number of times to run the benchmark')
+    parser.add_argument('-n', '--size', type=int, default=10**8, help='Problem size: length of array a')
+    parser.add_argument('-t', '--trials', type=int, default=3, help='Number of times to run the benchmark')
     parser.add_argument('--correctness-only', default=False, action='store_true', help='Only check correctness, not performance.')
     return parser